etl/pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = ["SetMao <set.mao.tw@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.35"
geoalchemy2 = "^0.15.2"
geopandas = "^1.0.1"
pandas = "^2.2.3"
requests-async = "^0.6.2"
ipython = "^8.28.0"
polars = "^1.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
